#fix: moved logic to helper function
